FROMLIST: LoongArch: Support mem=SIZE kernel parameter
The LoongArch mem= parameter parser was previously limited to the
mem=SIZE@START format. This was inconvenient for the common use case
of simply capping the total system memory, as it forced users to
manually specify a start address. It was also inconsistent with the
behavior on other architectures.

This patch enhances the parser in early_parse_mem() to also support the
more user-friendly mem=SIZE format. The implementation now checks for
the presence of the '@' symbol to determine the user's intent:

- If mem=SIZE is provided (no '@'), the kernel now calls
  memblock_enforce_memory_limit(). This trims memory from the top down
  to the specified size.
- If mem=SIZE@START is used, the original behavior is retained for
  backward compatibility. This allows for defining specific memory
  banks.

This change introduces an important usage rule reflected in the code's
comments: the mem=SIZE format should only be specified once on the
kernel command line. It acts as a single, global cap on total memory. In
contrast, the mem=SIZE@START format can be used multiple times to
define several distinct memory regions.

Signed-off-by: Ming Wang <wangming01@loongson.cn>

Link: https://lore.kernel.org/loongarch/20250701090449.2426151-1-wangming01@loongson.cn/
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>