deepin: firmware_loader: move log failed to load firmware to real failed
deepin inclusion

If we use the zstd compress firmware, cause the many report that
said what the firmware load failed, but the truth is that the load failed log
is not real.

Move it to the end of  fw_get_filesystem_firmware to check the real failed.

Link: https://bbs.deepin.org/zh/post/289309
Fixes: f081e61f3f1b ("debian: firmware_loader: Log direct loading failures as info for d-i")
Signed-off-by: Wentao Guan <guanwentao@uniontech.com>
(cherry picked from commit e3ba77c964c394c822c69656e2fb5c4b3fd33eeb)