FROMLIST: genirq: Add irq_chip_(startup/shutdown)_parent()
As the MSI controller on SG2044 uses PLIC as the underlying interrupt
controller, it needs to call the irq_enable() and irq_disable() to
startup/shutdown irqs. Otherwise, the MSI interrupt can not be startup
correctly and will not respond any incoming interrupt.

Introduce helper irq_chip_startup_parent() and irq_chip_shutdown_parent()
to allow the interrupt controller to call the irq_startup() or
irq_shutdown() of the parent interrupt chip. In case irq_startup() or
irq_shutdown() is not implemented for the parent interrupt chip, which
will fallback to irq_chip_enable_parent() or irq_chip_disable_parent().

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Inochi Amaoto <inochiama@gmail.com>
Link: https://lore.kernel.org/r/20250813232835.43458-2-inochiama@gmail.com
Signed-off-by: Han Gao <rabenda.cn@gmail.com>